Warnings removed. Divided into cloud and local. Added test Remove member
